[libc++] Use the internal Lit shell to run the tests
This makes the libc++ tests more portable -- almost all of them should
now work on Windows, except for some tests that assume a shell is
available on the target. We should probably provide a way to exclude
those anyway for the purpose of running tests on embedded targets.

Differential Revision: https://reviews.llvm.org/D89495

NOKEYCHECK=True
GitOrigin-RevId: 39bbfb77264a4a7a216921c2b70a30ba0f27eb56