[libc++] Fix broken modules tests on single-threaded systems
Since c0cd106fcc9f, we add __config_site macro defines to the compiler
command line whether we are building with modules or not. This means
that the modules tests are expected to fail on single-threaded systems
whether we build with modules or not.

Cr-Mirrored-From: https://chromium.googlesource.com/external/github.com/llvm/llvm-project
Cr-Mirrored-Commit: c82f9eba4a559e23bc0c356ac8fc0c3bda2801a1